pyproject.toml
[tool.poetry]
name = "multi-assets-sim"
version = "0.1.0"
description = "相関を持つ複数アセットでモンテカルロシミュレーションを行うGUIソフトウェア"
authors = ["ksk-log <150491853+ksk-log@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.1"
flet = "^0.10.3"
pandas = "^2.1.2"
matplotlib = "^3.8.0"
pyyaml = "^6.0.1"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
